+++ Sources/NowPlayingCore/PreferenceDefaults.swift
@@ -22,5 +22,5 @@ public enum PreferenceDefaults {
     public static let trackTemplate = "<artists> - <title> <(year)>"
 
     public static let popupOpacity: Double = 1
-    public static let minPopupOpacity: Double = 0.5
+    public static let minPopupOpacity: Double = 0.2
 }